requirements.txt
# Visão Computacional
opencv-python>=4.8.0
mediapipe>=0.10.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0
numpy>=1.24.0
scipy>=1.10.0

# Processamento de Dados
pandas>=2.0.0

# Comunicação
websockets>=11.0
python-socketio>=5.8.0

# Configuração
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilitários
loguru>=0.7.0

# Desenvolvimento (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
mypy>=1.5.0
